+++ tests/datasets/test_model_backwards_compatibility.py
@@ -5,17 +5,12 @@ from pathlib import Path
 
 import pytest
 
-from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
-    UnknownModelVersionError,
-)
-from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
-    add_container,
-)
-from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
-    convert_is_personal_data,
-)
-from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
-    copy_pseudonymization_metadata,
+from dapla_metadata.datasets.compatibility._utils import UnknownModelVersionError
+from dapla_metadata.datasets.compatibility._utils import add_container
+from dapla_metadata.datasets.compatibility._utils import convert_is_personal_data
+from dapla_metadata.datasets.compatibility._utils import copy_pseudonymization_metadata
+from dapla_metadata.datasets.compatibility._utils import (
+    is_metadata_in_container_structure,
 )
 from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
     handle_version_2_2_0,
@@ -28,9 +23,6 @@ from dapla_metadata.datasets.compatibility.model_backwards_compatibility import 
 )
 from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
     handle_version_5_0_1,
-)
-from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
-    is_metadata_in_container_structure,
 )
 from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
     upgrade_metadata,
